report update

+++ src/main/java/com/example/techgicus_ebilling/techgicus_ebilling/datamodel/entity/PurchaseReturn.java
@@ -30,7 +30,7 @@ public class PurchaseReturn {
 
     private Double totalTaxAmount;
 
-    private Integer totalQuantity;
+    private Double totalQuantity;
 
     private Double totalAmount;
 
@@ -122,11 +122,11 @@ public class PurchaseReturn {
         this.totalTaxAmount = totalTaxAmount;
     }
 
-    public Integer getTotalQuantity() {
+    public Double getTotalQuantity() {
         return totalQuantity;
     }
 
-    public void setTotalQuantity(Integer totalQuantity) {
+    public void setTotalQuantity(Double totalQuantity) {
         this.totalQuantity = totalQuantity;
     }
