+++ src/main/java/com/example/techgicus_ebilling/techgicus_ebilling/dto/saleReturnDto/SaleReturnItemResponse.java
@@ -12,7 +12,7 @@ public class SaleReturnItemResponse {
 
     private String name;
 
-    private Integer quantity;
+    private Double quantity;
 
     private Unit unit;
 
@@ -44,11 +44,11 @@ public class SaleReturnItemResponse {
         this.name = name;
     }
 
-    public Integer getQuantity() {
+    public Double getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Integer quantity) {
+    public void setQuantity(Double quantity) {
         this.quantity = quantity;
     }
